Move WooPosIsWooCommerceVersionSunsetWarningRequired to localcatalog package
